+++ src/main/java/ru/vyarus/guice/persist/orient/repository/core/executor/impl/GraphRepositoryExecutor.java
@@ -1,6 +1,5 @@
 package ru.vyarus.guice.persist.orient.repository.core.executor.impl;
 
-import com.google.common.collect.ImmutableList;
 import com.google.inject.Provider;
 import com.orientechnologies.orient.core.command.OCommandRequest;
 import com.tinkerpop.blueprints.Edge;
@@ -11,7 +10,6 @@ import ru.vyarus.guice.persist.orient.repository.core.executor.RepositoryExecuto
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
-import java.util.List;
 
 /**
  * Graph connection repository executor.
@@ -22,7 +20,6 @@ import java.util.List;
 @Singleton
 public class GraphRepositoryExecutor implements RepositoryExecutor {
 
-    private static final List<Class> ACCEPT_TYPES = ImmutableList.<Class>of(Vertex.class, Edge.class);
     private final Provider<OrientBaseGraph> provider;
 
     @Inject
@@ -32,7 +29,7 @@ public class GraphRepositoryExecutor implements RepositoryExecutor {
 
     @Override
     public boolean accept(final Class<?> returnType) {
-        return ACCEPT_TYPES.contains(returnType);
+        return Edge.class.isAssignableFrom(returnType) || Vertex.class.isAssignableFrom(returnType);
     }
 
     @Override